Support for user supplied error function
Defines a std::function for errors that takes an error string for
handling errors.  Default behavior is to print the error to stderr and
exit.

Functions that call the error function also return a boolean indicating
success or failure.  This allows the caller to handle exceptional cases
if the error function doesn't exit.