Add range parsing
Add range parsing, which takes a comma delimited list of numbers
or an inclusive range with an ellipsis.

Inputs of 1,2,3,4,5 and 1...5 will both result in a vector that is
equivalent to { 1, 2, 3, 4, 5 }.

The comma delimited input can be used for specific values without
including all numbers. For example, 1,3,5 will skip 2 and 4.

The ellipsis input will increment by 1 in the data type specified
by the vector argument.